trackable/agents/chatbot.py
"""
Vanilla chatbot agent for Ingress API.

This is a basic agent that uses Google ADK to chat with users via Gemini models.
Additional capabilities will be added as other components become ready.
"""

from google.adk.agents.llm_agent import Agent

# Create the vanilla chatbot agent
chatbot_agent = Agent(
    model="gemini-2.5-flash",
    name="trackable_chatbot",
    description="Personal shopping assistant for post-purchase order management",
    instruction="""You are Trackable, a helpful personal shopping assistant.

Your role is to help users manage their online shopping orders after purchase.

Current capabilities:
- Answer general questions about order management
- Provide friendly, helpful conversation
- Explain what Trackable can do once fully configured

Future capabilities (coming soon):
- Track orders from confirmation emails
- Monitor return and exchange windows
- Understand merchant policies
- Proactively remind users of approaching deadlines
- Recommend actions (return, exchange, or keep items)

For now, be friendly and helpful. Let users know that full order tracking capabilities
are being set up and will be available soon.
""",
)


__all__ = ["chatbot_agent"]
