pyproject.toml
[project]
description = "Trackable: Personal Shopping Agent"
name = "trackable"
readme = "README.md"
requires-python = ">=3.14"
version = "0.1.0"

dependencies = [
  "beautifulsoup4>=4.12.0",
  "cloud-sql-python-connector[pg8000]>=1.20.0",
  "dotenv>=0.9.9",
  "email-validator>=2.3.0",
  "fastapi>=0.123.10",
  "google-adk>=1.22.1",
  "google-api-python-client>=2.154.0",
  "google-auth-httplib2>=0.2.0",
  "google-auth-oauthlib>=1.2.0",
  "httpx>=0.28.1",
  "pillow>=11.2.0",
  "pydantic>=2.12.5",
  "sqlalchemy>=2.0.45",
  "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
  "black>=26.1.0",
  "isort>=7.0.0",
  "pytest>=9.0.2",
  "pytest-asyncio>=1.3.0",
  "pytest-cov>=7.0.0",
  "pytest-datadir>=1.8.0",
]

[tool.pytest]
addopts = ["-v", "-m", "not manual"]
markers = [
  "manual: marks tests as manual (deselected by default)",
]
minversion = "9.0"
testpaths = ["tests"]

[tool.isort]
profile = "black"
skip_gitignore = true
